requirements.txt
numpy==1.26.4
pillow==10.4.0
scikit-learn==1.5.2
matplotlib==3.9.2
imageio==2.35.1
